Move FakeIO to its own module to mirror SerialPort and ZMQ_IO
