+++ docs/api/fakeio.rst
@@ -1,0 +1,7 @@
+FakeIO IO Interface
+---------------------
+
+.. automodule:: larpix.fakeio
+    :members:
+    :undoc-members:
+
